layout component for backoffice
