--- src/Components/Layout/LayoutBackOffice.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import Sidebar from "../Sidebar/Sidebar"
-
-function LayoutBackOffice() {
-  return (
-      <Sidebar />
-    )
-}
-
-export default LayoutBackOffice
